telegram_excel_bot/requirements.txt
python-telegram-bot==21.6
openpyxl==3.1.5
filelock==3.16.1
python-dotenv==1.0.1
openai==1.54.4
httpx==0.27.2